remove duplicate fixture
